// modules/apps/headless/headless-admin-site/headless-admin-site-client-js/src/models/FragmentViewportStyle.ts
/**
 * SPDX-FileCopyrightText: (c) 2026 Liferay, Inc. https://liferay.com
 * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
 */


/**
 * @author Rubén Pulido
 * @generated
 */

	/**
	* The fragment's viewport style.
	*/
	export class FragmentViewportStyle {
			"backgroundColor"?: string;
			"borderColor"?: string;
			"borderRadius"?: string;
			"borderWidth"?: string;
			"fontFamily"?: string;
			"fontSize"?: string;
			"fontWeight"?: string;
			"height"?: string;
			"hidden"?: boolean;
			"marginBottom"?: string;
			"marginLeft"?: string;
			"marginRight"?: string;
			"marginTop"?: string;
			"maxHeight"?: string;
			"maxWidth"?: string;
			"minHeight"?: string;
			"minWidth"?: string;
			"opacity"?: string;
			"overflow"?: string;
			"paddingBottom"?: string;
			"paddingLeft"?: string;
			"paddingRight"?: string;
			"paddingTop"?: string;
			"shadow"?: string;
			"textAlign"?: string;
			"textColor"?: string;
			"width"?: string;

		static "discriminator": string | undefined = undefined;

	static "attributeTypeMap": Array<{
		baseName: string;
		name: string;
		type: string;
	}> = [
		{
			baseName: "backgroundColor",
			name: "backgroundColor",
			type: "string",
		},
		{
			baseName: "borderColor",
			name: "borderColor",
			type: "string",
		},
		{
			baseName: "borderRadius",
			name: "borderRadius",
			type: "string",
		},
		{
			baseName: "borderWidth",
			name: "borderWidth",
			type: "string",
		},
		{
			baseName: "fontFamily",
			name: "fontFamily",
			type: "string",
		},
		{
			baseName: "fontSize",
			name: "fontSize",
			type: "string",
		},
		{
			baseName: "fontWeight",
			name: "fontWeight",
			type: "string",
		},
		{
			baseName: "height",
			name: "height",
			type: "string",
		},
		{
			baseName: "hidden",
			name: "hidden",
			type: "boolean",
		},
		{
			baseName: "marginBottom",
			name: "marginBottom",
			type: "string",
		},
		{
			baseName: "marginLeft",
			name: "marginLeft",
			type: "string",
		},
		{
			baseName: "marginRight",
			name: "marginRight",
			type: "string",
		},
		{
			baseName: "marginTop",
			name: "marginTop",
			type: "string",
		},
		{
			baseName: "maxHeight",
			name: "maxHeight",
			type: "string",
		},
		{
			baseName: "maxWidth",
			name: "maxWidth",
			type: "string",
		},
		{
			baseName: "minHeight",
			name: "minHeight",
			type: "string",
		},
		{
			baseName: "minWidth",
			name: "minWidth",
			type: "string",
		},
		{
			baseName: "opacity",
			name: "opacity",
			type: "string",
		},
		{
			baseName: "overflow",
			name: "overflow",
			type: "string",
		},
		{
			baseName: "paddingBottom",
			name: "paddingBottom",
			type: "string",
		},
		{
			baseName: "paddingLeft",
			name: "paddingLeft",
			type: "string",
		},
		{
			baseName: "paddingRight",
			name: "paddingRight",
			type: "string",
		},
		{
			baseName: "paddingTop",
			name: "paddingTop",
			type: "string",
		},
		{
			baseName: "shadow",
			name: "shadow",
			type: "string",
		},
		{
			baseName: "textAlign",
			name: "textAlign",
			type: "string",
		},
		{
			baseName: "textColor",
			name: "textColor",
			type: "string",
		},
		{
			baseName: "width",
			name: "width",
			type: "string",
		},
		];

		static getAttributeTypeMap() {
				return FragmentViewportStyle.attributeTypeMap;
		}
	}